refactor: league 패키지 생성 및 IntegrationTest 생성
