+++ soccer-api/src/main/kotlin/com/example/soccerapi/game/application/LeagueGameFacade.kt
@@ -1,7 +1,7 @@
 package com.example.soccerapi.game.application
 
-import com.example.soccerdomain.game.LeagueGameFactory
-import com.example.soccerdomain.game.LeagueGameRepository
+import com.example.soccerdomain.game.league.LeagueGameFactory
+import com.example.soccerdomain.game.league.LeagueGameRepository
 import com.example.soccerdomain.game.season.Season
 import com.example.soccerdomain.game.record.TeamRecord
 import com.example.soccerdomain.game.record.TeamRecordRepository
