+++ soccer-api/src/test/kotlin/com/example/soccerapi/game/application/LeagueGameFacadeTest.kt
@@ -1,30 +1,19 @@
 package com.example.soccerapi.game.application
 
-import com.example.soccerdomain.game.LeagueGameRepository
-import com.example.soccerdomain.game.record.TeamRecordRepository
-import com.example.soccerdomain.game.season.SeasonRepository
+import com.example.soccerapi.support.ApiIntegrationTest
 import com.example.soccerdomain.team.domain.League
 import com.example.soccerdomain.team.domain.Team
-import com.example.soccerdomain.team.domain.TeamRepository
-import jakarta.persistence.EntityManager
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatCode
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.transaction.annotation.Transactional
 
-@SpringBootTest
 @Transactional
 class LeagueGameFacadeTest @Autowired constructor(
-    val leagueGameFacade: LeagueGameFacade,
-    val leagueGameRepository: LeagueGameRepository,
-    val teamRecordRepository: TeamRecordRepository,
-    val teamRepository: TeamRepository,
-    val seasonRepository: SeasonRepository,
-    val entityManager: EntityManager
-) {
+    val leagueGameFacade: LeagueGameFacade
+) : ApiIntegrationTest() {
 
 
     @Test
