+++ soccer-domain/src/main/kotlin/com/example/soccerdomain/game/league/LeagueGameFactory.kt
@@ -1,4 +1,4 @@
-package com.example.soccerdomain.game
+package com.example.soccerdomain.game.league
 
 import com.example.soccerdomain.team.domain.League
 import com.example.soccerdomain.team.domain.Team
