+++ soccer-domain/src/test/kotlin/com/example/soccerdomain/player/domain/PlayerCustomRepositoryImplTestDomain.kt
@@ -1,28 +1,19 @@
 package com.example.soccerdomain.player.domain
 
-import com.example.soccerdomain.SoccerConfig
 import com.example.soccerdomain.common.search.NumberOperator
 import com.example.soccerdomain.common.search.OrderBy
 import com.example.soccerdomain.player.NumberSearchInput
 import com.example.soccerdomain.player.PlayerSearchInput
+import com.example.soccerdomain.support.DomainIntegrationTest
 import com.example.soccerdomain.team.domain.League
 import com.example.soccerdomain.team.domain.Team
-import com.example.soccerdomain.team.domain.TeamRepository
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration
-import org.springframework.boot.test.context.SpringBootTest
 
 
-@EnableAutoConfiguration
-@SpringBootTest(classes = [SoccerConfig::class])
-class PlayerCustomRepositoryImplTest @Autowired constructor(
-    val playerRepository: PlayerRepository,
-    val teamRepository: TeamRepository
-) {
+class PlayerCustomRepositoryImplTestDomain : DomainIntegrationTest() {
 
     @AfterEach
     fun tearDown() {
